+++ packages/analytics-react-native/test/react-native-client.test.ts
@@ -1,12 +1,16 @@
 import { AmplitudeReactNative } from '../src/react-native-client';
 import * as core from '@amplitude/analytics-core';
 import * as CookieMigration from '../src/cookie-migration';
-import { Status, UserSession, Event } from '@amplitude/analytics-types';
+import {
+  Status,
+  UserSession,
+  Event,
+  getAnalyticsConnector,
+  getCookieName as getStorageKey,
+} from '@amplitude/analytics-core';
 import { isWeb } from '../src/utils/platform';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import { getAnalyticsConnector } from '@amplitude/analytics-client-common';
 import * as Config from '../src/config';
-import { getCookieName as getStorageKey } from '@amplitude/analytics-client-common/src';
 
 describe('react-native-client', () => {
   const API_KEY = 'API_KEY';
@@ -79,13 +83,13 @@ describe('react-native-client', () => {
       await Promise.all([
         client.init(API_KEY, USER_ID, {
           ...attributionConfig,
-        }),
+        }).promise,
         client.init(API_KEY, USER_ID, {
           ...attributionConfig,
-        }),
+        }).promise,
         client.init(API_KEY, USER_ID, {
           ...attributionConfig,
-        }),
+        }).promise,
       ]);
       // NOTE: `parseOldCookies` and `useNodeConfig` are only called once despite multiple init calls
       expect(parseOldCookies).toHaveBeenCalledTimes(1);
@@ -303,9 +307,7 @@ describe('react-native-client', () => {
   describe('reset', () => {
     test('should reset user id and generate new device id config', async () => {
       const client = new AmplitudeReactNative();
-      await client.init(API_KEY, undefined, {
-        ...attributionConfig,
-      }).promise;
+      await client.init(API_KEY).promise;
       client.setUserId(USER_ID);
       client.setDeviceId(DEVICE_ID);
       expect(client.getUserId()).toBe(USER_ID);
@@ -313,7 +315,7 @@ describe('react-native-client', () => {
       client.reset();
       expect(client.getUserId()).toBe(undefined);
       expect(client.getDeviceId()).not.toBe(DEVICE_ID);
-    });
+    }, 10000);
   });
 
   describe('getSessionId', () => {
@@ -464,7 +466,8 @@ describe('react-native-client', () => {
         this.handleAppStateChange('active');
       }
 
-      setBackground() {
+      setBackground(timestamp: number) {
+        this.currentTime = timestamp;
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
         this.handleAppStateChange('background');
@@ -502,46 +505,46 @@ describe('react-native-client', () => {
       const send = jest.fn().mockReturnValue(sendResponse);
       const cookieStorage = new core.MemoryStorage<UserSession>();
 
-      const client1 = new AmplitudeReactNativeTest(500);
+      const client1 = new AmplitudeReactNativeTest(950);
       await client1.init(API_KEY, undefined, clientOptions(send, cookieStorage, true)).promise;
 
       client1.setActive(1000);
 
-      expect(client1.config.sessionId).toEqual(1000);
+      expect(client1.config.sessionId).toEqual(950);
       expect(client1.config.lastEventTime).toEqual(1000);
-      expect(client1.config.lastEventId).toEqual(2);
+      expect(client1.config.lastEventId).toEqual(1);
 
       void client1.track({ event_type: 'event-1', time: 1200 });
 
-      expect(client1.config.sessionId).toEqual(1000);
+      expect(client1.config.sessionId).toEqual(950);
       expect(client1.config.lastEventTime).toEqual(1200);
-      expect(client1.config.lastEventId).toEqual(3);
+      expect(client1.config.lastEventId).toEqual(2);
 
       const client2 = new AmplitudeReactNativeTest(1250);
       await client2.init(API_KEY, undefined, clientOptions(send, cookieStorage, true)).promise;
 
-      expect(client2.config.sessionId).toEqual(1000);
+      expect(client2.config.sessionId).toEqual(950);
       expect(client2.config.lastEventTime).toEqual(1250);
-      expect(client2.config.lastEventId).toEqual(3);
+      expect(client2.config.lastEventId).toEqual(2);
 
       void client2.track({ event_type: 'event-2', time: 1270 });
 
-      expect(client2.config.sessionId).toEqual(1000);
+      expect(client2.config.sessionId).toEqual(950);
       expect(client2.config.lastEventTime).toEqual(1270);
-      expect(client2.config.lastEventId).toEqual(4);
+      expect(client2.config.lastEventId).toEqual(3);
 
       const client3 = new AmplitudeReactNativeTest(1300);
       await client3.init(API_KEY, undefined, clientOptions(send, cookieStorage, true)).promise;
 
-      expect(client3.config.sessionId).toEqual(1000);
+      expect(client3.config.sessionId).toEqual(950);
       expect(client3.config.lastEventTime).toEqual(1300);
-      expect(client3.config.lastEventId).toEqual(4);
+      expect(client3.config.lastEventId).toEqual(3);
 
       client3.setActive(1500);
 
       expect(client3.config.sessionId).toEqual(1500);
       expect(client3.config.lastEventTime).toEqual(1500);
-      expect(client3.config.lastEventId).toEqual(6);
+      expect(client3.config.lastEventId).toEqual(5);
     });
 
     describe('track session events', () => {
@@ -562,7 +565,7 @@ describe('react-native-client', () => {
 
         void client.track({ event_type: 'event-5', time: 1700 });
 
-        client.setBackground();
+        client.setBackground(1730);
 
         void client.track({ event_type: 'event-6', time: 1750 });
         void client.track({ event_type: 'event-7', time: 2000 });
@@ -574,7 +577,7 @@ describe('react-native-client', () => {
         // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
         const events = send.mock.calls.flatMap((call) => call[1].events as Event[]);
         expect(events.length).toEqual(18);
-        events.forEach((event, i) => expect(event.event_id).toEqual(i));
+        events.forEach((event, i) => expect(event.event_id).toEqual(i + 1));
 
         expect(events[0].event_type).toEqual('session_end');
         expect(events[0].session_id).toEqual(500);
@@ -673,7 +676,7 @@ describe('react-native-client', () => {
         client.currentTime = 1720;
         client.setSessionId(5050);
 
-        client.setBackground();
+        client.setBackground(1730);
 
         void client.track({ event_type: 'event-6', time: 1750 });
         void client.track({ event_type: 'event-7', time: 2000 });
@@ -687,7 +690,7 @@ describe('react-native-client', () => {
         // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
         const events = send.mock.calls.flatMap((call) => call[1].events as Event[]);
         expect(events.length).toEqual(14);
-        events.forEach((event, i) => expect(event.event_id).toEqual(i));
+        events.forEach((event, i) => expect(event.event_id).toEqual(i + 1));
 
         expect(events[0].event_type).toEqual('session_end');
         expect(events[0].session_id).toEqual(500);
@@ -765,7 +768,7 @@ describe('react-native-client', () => {
 
         void client.track({ event_type: 'event-5', time: 1700 });
 
-        client.setBackground();
+        client.setBackground(1730);
 
         void client.track({ event_type: 'event-6', time: 1750 });
         void client.track({ event_type: 'event-7', time: 2000 });
@@ -777,7 +780,7 @@ describe('react-native-client', () => {
         // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
         const events = send.mock.calls.flatMap((call) => call[1].events as Event[]);
         expect(events.length).toEqual(8);
-        events.forEach((event, i) => expect(event.event_id).toEqual(i));
+        events.forEach((event, i) => expect(event.event_id).toEqual(i + 1));
 
         expect(events[0].event_type).toEqual('event-1');
         expect(events[0].session_id).toEqual(950);
@@ -836,7 +839,7 @@ describe('react-native-client', () => {
         client.currentTime = 1720;
         client.setSessionId(5050);
 
-        client.setBackground();
+        client.setBackground(1730);
 
         void client.track({ event_type: 'event-6', time: 1750 });
         void client.track({ event_type: 'event-7', time: 2000 });
@@ -850,7 +853,7 @@ describe('react-native-client', () => {
         // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
         const events = send.mock.calls.flatMap((call) => call[1].events as Event[]);
         expect(events.length).toEqual(8);
-        events.forEach((event, i) => expect(event.event_id).toEqual(i));
+        events.forEach((event, i) => expect(event.event_id).toEqual(i + 1));
 
         expect(events[0].event_type).toEqual('event-1');
         expect(events[0].session_id).toEqual(1000);