+++ packages/plugin-global-user-properties/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.70](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-global-user-properties@1.2.69...@amplitude/plugin-global-user-properties@1.2.70) (2025-07-18)
+
+**Note:** Version bump only for package @amplitude/plugin-global-user-properties
+
+
+
+
+
+## [1.2.69](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-global-user-properties@1.2.68...@amplitude/plugin-global-user-properties@1.2.69) (2025-07-17)
+
+**Note:** Version bump only for package @amplitude/plugin-global-user-properties
+
+
+
+
+
 ## [1.2.68](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-global-user-properties@1.2.67...@amplitude/plugin-global-user-properties@1.2.68) (2025-07-15)
 
 **Note:** Version bump only for package @amplitude/plugin-global-user-properties